Add a few links (#50)
* Add OpenAV

There are several kinds of software in OpenAV. It would be a little bit
messy to put them into catogories seperately. So I list them as a
collection.

* Add Sonic Pi

Although it's about coding, it's meant to be used as an live performance
app/instrument. So I add it to Audio apps.

* Add Bespoke

You may have seen this on twitter. It's a nice modular synth to
play with. Also the author quite active on the project.