Update plugins.json
NOTE: when the project is reloaded, this VST will crash. make sure to click on it so it can "wake up"
VST needs an output node to play sound, only spawn one or you guessed it, the program will crash!
Midi nodes and sound nodes will also crash the VST.
plus a midi note can hang, you have to restart your DAW if that happens

These issues will be fixed over time, but I just added VST support and the program is very capable and expansive.